✨ Set initial footer props at compile time
Both the server and client configurations are important here. The server
one isn't sufficient on its own, but it's key to ensuring that the
replaced values are present in the output source html - without it the
replacment is done by the client and there's a flash.

This is true even when the replacement targets are placed directly in
the html, or in the js - as I'm doing here, because it's neater and
doesn't seem to make a difference.