💄 Make CSS spacing more sensible
Shed a tear for my 'wizardry' calc: this was meant to decrease line
height slightly as font size scaled up. It seemed to work ok at my
screen sizes, but would have needed media queries to keep it in check,
and felt like a generally bad idea.

Here I'm replacing that with a static decimal representation of the
golden ratio, and also adding a set of modularly scaled spacer vars.