🐛 Fix header styling
Sorting out some FF/Chrome discrepancies here:
The idea is that that header logotype should grow to be quite large, but
stop before it gets ginormous.

In FF it did just fine with only the padding and max-width, but in
Chrome it was shy about growing until I added column template.

The vertical whitespace above and below the logotype should grow and
shrink with the viewport width, but should not be affected by the
content below it (i.e. when switching pages). Adjusting the row to
min-content and using padding instead takes care of this.