🐛 Fix flash of too-long line measure
This includes the 0 character in the subset of the serif font - used
until the js at the bottom of the page loads the full set.

The 'ch' unit I'm now using to regulate column width is based on this
character, so it's become critical to the layout. I could have just
switched the measure from 70ch to 35em and saved the 2KB, but eh...

... yeah I probably should have done that.