🏗 Set keychain-aquired API key as environment var
Instead of returning a promise which resolves to the API key, this
script now does something completely different:

Takes the name of an npm script as argument, then spawns that script in
a child process where the API key has been made available.

This allows me to continue storing my secrets used by the dev
environment in my OS keychain (as opposed to plaintext .env), without
needing to require keytar or dotenv inside any scripts that become part
of the webpack bundle.

That, in turn, eliminates the need for me to write consumed API data to
a file, opening an opportunity to make API calls more directly.