✨ Decorated Hex Color Codes
This adds an inline markdown-it plugin to find hex colors in content,
wrap them in a css class, and prepend them with a colored dot.

I'm not sure I understand the markdown-it API well enough to know if I'm
doing this "right", but it works and it doesn't seem to break anything!